chore(adv): checkpoint tk-9ad1a04909a2
Change: addAdvMcpReadSurface
Task: tk-9ad1a04909a2
Mode: complete
Verification: SDK-free catalog module extracted. Red run tr_mru2eaor_ef3725a6 (8 tests failed — module didn't exist); GREEN run tr_mru2ol3c_23a4d3e6 — 51/51 tests pass across tool-catalog-entries.test.ts + tool-registry.inventory.test.ts + tool-catalog.test.ts. Verify run tr_mru2ozur_b22836a2 — 62/62 tests pass across tool-registry.test.ts + tools/contract.test.ts + tools/adv-invoke.test.ts (no regressions). Typecheck green. Files: NEW plugin/src/tool-catalog-entries.ts (SDK-free types/functions/constants), NEW plugin/src/tool-catalog-entries.test.ts (import-deps + parity), MODIFY plugin/src/tool-registry.ts (re-export + import from new module; PUBLIC_TOOL_ENTRIES now export const; getToolSurface wrapped for backward compat). KD2 deviation: PUBLIC_TOOL_GROUPS data remains in tool-registry.ts (SDK-coupled via tool modules); MCP will use dynamic import for runtime data (same pattern as adv_contract_mint).
